dispatcher/requirements.txt
google-api-python-client==2.87.0
flask==2.2.5
functions-framework==3.3.0
slackclient==2.9.4
requests==2.31.0
google-cloud-pubsub==2.18.0